requirements.txt
clipboard~=0.0.4
PyYAML~=6.0.1
tk~=0.1.0
clipboard~=0.0.4